+++ src/features/migration/Migration.types.ts
@@ -73,6 +73,7 @@ export interface MigrationBulkSearchSettings {
     selectHighestChapterNumberSource: boolean;
     ignoreOutdatedMatches: boolean;
     requireAdditionalChapters: boolean;
+    ignoreWithMissingChapters: boolean;
     performAdvancedSearch: boolean;
 }
 
@@ -102,6 +103,7 @@ export interface MigrationMatch
     extends MangaIdInfo, MangaTitleInfo, MangaThumbnailInfo, MangaSourceIdInfo, MangaArtistInfo, MangaAuthorInfo {
     sourceTitle: SourceDisplayNameInfo['displayName'] | undefined;
     latestChapterNumber: ChapterNumberInfo['chapterNumber'] | undefined;
+    missingChapters: number | undefined;
 }
 
 export interface TMigrationEntry {
@@ -110,6 +112,7 @@ export interface TMigrationEntry {
     mangaArtist: MangaArtistInfo['artist'];
     mangaAuthor: MangaAuthorInfo['author'];
     latestChapterNumber: ChapterNumberInfo['chapterNumber'] | undefined;
+    missingChapters: number | undefined;
     mangaThumbnailUrl: MangaThumbnailInfo['thumbnailUrl'] | undefined;
     sourceId: SourceIdInfo['id'];
     sourceTitle: SourceDisplayNameInfo['displayName'] | undefined;